Resolvendo conflitos com a develop
